+++ packages/core/src/types/form.qwik.ts
@@ -10,8 +10,8 @@ import type { DeepPartial } from './utils.ts';
 export interface FormConfig<TSchema extends Schema = Schema> {
   readonly schema: TSchema;
   readonly initialInput?: DeepPartial<v.InferInput<TSchema>> | undefined;
-  readonly validateOn?: ValidationMode | undefined;
-  readonly revalidateOn?: Exclude<ValidationMode, 'initial'> | undefined;
+  readonly validate?: ValidationMode | undefined;
+  readonly revalidate?: Exclude<ValidationMode, 'initial'> | undefined;
 }
 
 export interface InternalFormStore<TSchema extends Schema = Schema>
@@ -19,9 +19,9 @@ export interface InternalFormStore<TSchema extends Schema = Schema>
   element?: HTMLFormElement;
 
   validators: number;
-  validateOn: ValidationMode;
-  revalidateOn: Exclude<ValidationMode, 'initial'>;
-  validate: QRL<(input: unknown) => Promise<v.SafeParseResult<TSchema>>>;
+  validate: ValidationMode;
+  revalidate: Exclude<ValidationMode, 'initial'>;
+  parse: QRL<(input: unknown) => Promise<v.SafeParseResult<TSchema>>>;
 
   isSubmitting: Signal<boolean>;
   isSubmitted: Signal<boolean>;