+++ frameworks/preact/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to the library will be documented in this file.
 
 - Add `SubmitHandler` type from `@formisch/core` to exports
 - Change `render` prop to `children` in `Field` and `FieldArray` component
+- Change name of `validateOn` and `revalidateOn` config by removing `On` suffix
+- Change name of `revalidateOn` config to `revalidate` in internal form store
 - Refactor `useField` and `useFieldArray` to improve implementation
 - Refactor `Form` component to use new `handleSubmit` method
 - Fix bug when resetting dynamic field array items