+++ frameworks/qwik/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Add `SubmitHandler` type from `@formisch/core` to exports
+- Change name of `validateOn` and `revalidateOn` config by removing `On` suffix
 - Refactor `useField` and `useFieldArray` to improve implementation
 - Refactor `Form` component to use new `handleSubmit` method
 - Fix bug when resetting dynamic field array items