+++ packages/core/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Fix bug in `resetItemState` by also resetting `.errors` signals
+- Change name of `validate` method to `parse` in internal form store
+- Change name of `validateOn` config to `validate` in internal form store
+- Change name of `revalidateOn` config to `revalidate` in internal form store
 
 ## v0.1.0 (July 14, 2025)